Refactored SideBar, improved styling on mobile
