+++ src/components/NavBar.tsx
@@ -1,9 +1,11 @@
+import { useLocation } from "@gatsbyjs/reach-router"
 import classNames from "classnames"
 import { graphql, Link, useStaticQuery } from "gatsby"
 import { useContext, useState } from "react"
 import { PageContext } from "../contexts/PageContext"
 import { StrapiNavigation } from "../types/strapi"
 import BurgerButton from "./BurgerButton"
+import NavPages from "./NavPages"
 
 type NavQuery = {
   allStrapiPublicNavigation: {
@@ -14,6 +16,7 @@ type NavQuery = {
 const NavBar: React.FC = () => {
   const { locale, localeLink } = useContext(PageContext)
   const [open, setOpen] = useState(false)
+  const { pathname } = useLocation()
   const navigations = useStaticQuery(graphql`
     query {
       allStrapiPublicNavigation {
@@ -56,7 +59,7 @@ const NavBar: React.FC = () => {
         />
         <div
           className={classNames(
-            "fixed top-0 left-0 h-full bg-gray-darkest transition-[max-width] ease-in-out duration-500",
+            "fixed top-0 left-0 h-full bg-gray-darkest transition-all ease-in-out duration-500 mt-12",
             {
               "max-w-[80%] w-full overflow-y-auto": open,
               "max-w-0 w-0 overflow-hidden": !open,
@@ -65,19 +68,22 @@ const NavBar: React.FC = () => {
         >
           <div className="p-4">
             {nav?.items.map((item) => (
-              <Link
-                key={item.path}
-                to={item.path}
-                className="block font-mono text-2xl p-4 text-white"
-                activeClassName="underline"
-                partiallyActive
-              >
-                {item.title}
-              </Link>
+              <>
+                <Link
+                  key={item.path}
+                  to={item.path}
+                  className="block font-mono text-2xl p-4 text-white"
+                  activeClassName="underline"
+                  partiallyActive
+                >
+                  {item.title}
+                </Link>
+                {pathname.includes(item.path) && <NavPages />}
+              </>
             ))}
             <Link
               to={localeLink}
-              className="block font-mono text-2xl p-4 text-white"
+              className="inline-block font-mono text-2xl p-4 text-white overflow-hidden w-60"
             >
               {locale === "fi" ? "In English" : "Suomeksi"}
             </Link>
